separate Apps and Files commands in different sub folders
