+++ src/App/Commands/Files/FilesUploadCommand.php
@@ -1,13 +1,14 @@
 <?php
 
 
-namespace Console\App\Commands;
+namespace Console\App\Commands\Files;
 
 
 use GuzzleHttp\Exception\GuzzleException;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
+use Console\App\Commands\Command;
 
 class FilesUploadCommand extends Command
 {
